Add `trace`-level logging via the `log` facade
We recently stumbled across some harder to debug issues that highlighted
that the absence of logging in `vss-client-ng` can be a pain. Therefore,
we here introduce logging for requests and the retry logic based on the
`log` facade, which not only allows the logging to easily integrate with
the Rust ecosystem, but is also a non-API breaking change allowing us to
ship this in a patch release.